+++ src/Resources/app/administration/src/module/sw-media/component/sw-filerobot-library/index.js
@@ -433,9 +433,7 @@ Component.register('sw-filerobot-library', {
                                                     media = await this.mediaRepository.get(media_id, Context.api);
                                                     if (media.uploadedAt !== null) {
                                                         //override image url in selection
-                                                        this.selection[this.selection.length - 1].url = selected.file.url.cdn;
-                                                        this.selectedItems = this.selection;
-                                                        
+
                                                         checkUpload = true;
                                                         let mediaURL = media.url;
                                                         let mediaPath = mediaURL.replace(domainUrl, '');
@@ -458,6 +456,12 @@ Component.register('sw-filerobot-library', {
                                                                 "media_path": '/' + mediaPath
                                                             }
                                                         )
+
+                                                        if (this.selection.length === 0) {
+                                                            this.selection.push(media);
+                                                        }
+                                                        this.selection[this.selection.length - 1].url = selected.file.url.cdn;
+                                                        this.selectedItems = this.selection;
                                                     }
                                                 }
                                             }
@@ -483,7 +487,7 @@ Component.register('sw-filerobot-library', {
                                 modalElement.querySelector('.sw-button.sw-button--primary').click();
                                 this.$emit('media-selection-change', this.selectedItems);
 
-                                await this.sleep(800);
+                                await this.sleep(1000);
                                 if (mediaArray.length) {
                                     for (let i = 0; i < mediaArray.length; i++) {
                                         let deleteURL = domainUrl + 'api/scaleflex/filerobot/clean-up-media';